First steps of attack rolling methods
Created atk_roll method on Weapon class
Created roll_damage method on Game_Master class
Envisioned damage structure and hierarchy
