+++ Study_Case_2.py
@@ -1,81 +1,6 @@
-import Study_Case_1 as sc_i
-import Global_Elements as globals
-from math import log10
-
-class Entity():
-    
-    def __init__(self) -> None:
-        """A collection of the fundamental parts for any entity.
-        
-        Current properties (and build patterns) include:
-         - name
-         - attr_values
-         - attr_die
-         
-         Current methods include:
-         - roll_stat: rolls for stat and handles the math, returning the power level
-         - battle_stats: returns a tuple with the entity's conflict stats
-         """
-        self.name = ''
-        self.attr_values = globals.build_attr_values_dict()
-        self.attr_die = globals.build_attr_dice_dict()
-
-    def set_attribute(self, attribute: str, value: int):
-        """Builder pattern for attr_values.
-        
-        The attribute names are as follows:
-         - STR: Strength
-         - DEX: Dexterity
-         - VIT: Vitality
-         - PER: Perception
-         - INT: Inteligence
-         - CHA: Charisma
-         - LCK: Luck
-         - WIS: Wisdom
-         - ARC: Arcane
-         - PAT: Patience
-         """
-        
-        if attribute.upper() not in globals.attributes:  # Error Handling
-            raise ValueError('Invalid Attribute')
-        if type(value) != int or value < 0:
-            raise TypeError('Attribute must be a non-negative integer')
-        
-        self.attr_values[attribute.upper()] = value
-        return self
-
-    def set_name(self, name: str):
-        "Builder pattern for name"
-
-        self.name = name
-        return self
-    
-
-    def battle_stats(self):
-        """Conflict stats generator. Returns a List with the following:
-        
-        - Given Entity
-        - Entity's HP: calculate through: 5**log10(VIT) * log10(VIT)*2
-        """
-
-        log_vit = log10(self.attr_values['VIT'])
-        health_points = int(5**log_vit * log_vit*2)
-
-        return (self, health_points)
-
-    
-    @classmethod
-    def roll_stat(cls, char, attribute: str, scale: int = 0.1):
-        "Rolls a given stat for a given character, according to it's corresponding dice (defaults to d20)."
-        if attribute.upper() not in globals.attributes:  # Error Handling
-            raise ValueError('Invalid Attribute')
-        
-        dice = char.attr_die[attribute.upper()]  # upper() method drops the case-sensitiveness
-        if dice is None:
-            dice = sc_i.Dice().set_num_sides(20)
-        
-        modifier = sc_i.Dice.roll(dice) * scale
-        return char.attr_values[attribute.upper()] * (modifier+0.4)
+from Entity_Class import Entity
+import Study_Case_3 as sc_iii
+import Global_Elements as gl
     
 
 class Game_Master():
@@ -122,9 +47,15 @@ class Game_Master():
         if roll_1 > roll_2:
             return (0, roll_1)
         return (1, roll_2)
+    
+    @classmethod 
+    def roll_damage(cls, entity: Entity, weapon:sc_iii.Weapon):
+        return sc_iii.Weapon.atk_roll(weapon)
 
 
 if __name__ == "__main__":
+    import Study_Case_1 as sc_i
+
     dave = Entity().set_attribute('Vit', 3156)
     joe = Entity().set_attribute('Str', 2865)
 
@@ -133,3 +64,6 @@ if __name__ == "__main__":
     print(game.clash_stats(dave, 'Vit', joe, 'Str'))
     game.delta_HP(dave, -50)
     print(game.active_entities[dave])
+
+    stick = sc_iii.Weapon().set_dmg_dice(sc_i.Dice().set_num_sides(4))
+    print(game.roll_damage(dave, stick))
